Move tests for common to cpp/tests/common.
