+++ cmake/metaspore_tests.cmake
@@ -53,13 +53,14 @@ if(BUILD_SERVING_BIN)
         add_test(NAME ${test_name} COMMAND ${test_name})
     endfunction()
 
-    add_cpp_test(test_threadpool serving/threadpool_test.cpp)
+    add_cpp_test(test_threadpool common/threadpool_test.cpp)
+    add_cpp_test(test_schema_parser common/schema_parser_test.cpp)
+    add_cpp_test(test_arrow_plan common/arrow_plan_test.cpp)
+    add_cpp_test(test_feature_compute_funcs common/feature_compute_funcs_test.cpp)
+    add_cpp_test(test_feature_compute_exec common/feature_compute_exec_test.cpp)
+
     add_cpp_test(test_ort_model serving/ort_model_test.cpp)
     add_cpp_test(test_sparse_lookup_model serving/sparse_lookup_model_test.cpp)
-    add_cpp_test(test_feature_compute_exec serving/feature_compute_exec_test.cpp)
-    add_cpp_test(test_feature_compute_funcs serving/feature_compute_funcs_test.cpp)
-    add_cpp_test(test_arrow_plan serving/arrow_plan_test.cpp)
-    add_cpp_test(test_schema_parser serving/schema_parser_test.cpp)
     add_cpp_test(test_sparse_ctr_model serving/sparse_ctr_model_test.cpp)
     add_cpp_test(test_tabular_xgboost_model serving/tabular_xgboost_test.cpp)
