+++ cpp/tests/common/feature_compute_funcs_test.cpp
@@ -15,7 +15,7 @@
 //
 
 #include <common/hash_utils.h>
-#include <serving/test_utils.h>
+#include <common/test_utils.h>
 #include <common/utils.h>
 
 #include <arrow/api.h>
