+++ cpp/tests/serving/sparse_ctr_model_test.cpp
@@ -19,7 +19,7 @@
 #include <serving/feature_extraction_model_input.h>
 #include <serving/ort_model.h>
 #include <serving/tabular_model.h>
-#include <serving/test_utils.h>
+#include <common/test_utils.h>
 #include <common/threadpool.h>
 #include <common/types.h>
