+++ cpp/tests/serving/sparse_lookup_model_test.cpp
@@ -17,7 +17,7 @@
 #include <arrow/api.h>
 #include <common/arrow/arrow_helpers.h>
 #include <serving/sparse_lookup_model.h>
-#include <serving/test_utils.h>
+#include <common/test_utils.h>
 
 #include <thread>
